Extract Location and LocationList components
I'm dropping the separate .css files for components. I find it easier to
reason about them when the styles are defined together with the
component; if I keep components as small as I intend to, and extract
style constants where appropriate, things should remain manageable even
as the styles get more complex.