Add lib support for discarding locations
Also introduced an integration test in lib for the main workflow. This
isn't an integration test over the whole system, as it does not exercise
the GUI, but it integrates over all of lib.

This will be faster to work with for trivial use-cases (like discarding)
than separate unit tests for each part of the loop, but I'm still happy
with the more focused tests for more complex bits like the
initialize-app action.